use consistent folder names for digital-twin dapps
- [CORE-599]
