+++ dapps/digital-twin-detail/src/components/DigitalTwinDetail.vue
