+++ dapps/digital-twin-detail/src/i18n/en.json
