+++ dapps/digital-twin-detail/src/routes.ts
