+++ dapps/digital-twin-detail/tsconfig.json
@@ -13,7 +13,7 @@
   ],
   "references": [
     {
-      "path": "../digital-twin.lib"
+      "path": "../digital-twin-lib"
     }
   ]
 }
