+++ dapps/digital-twin-lib/src/i18n/de.json
