+++ dapps/digital-twin-lib/src/i18n/index.ts
